committed removed Double file of text editor test cases
